requirements.txt
discord.py
requests
tweepy
